Codebase Refactor + Layout Changes + Deleted RootLayout
