+++ package.json
@@ -29,6 +29,7 @@
     "clsx": "^2.0.0",
     "cuid": "^3.0.0",
     "dotenv": "^16.3.1",
+    "framer-motion": "^10.18.0",
     "lucide-react": "^0.303.0",
     "next": "14.0.4",
     "next-auth": "^4.24.5",
