+++ src/app/(content)/layout.tsx
@@ -1,11 +1,11 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
-import "./globals.css";
-import NavBar from "./components/NavBar";
-import Provider from "./context/provider";
+import "../globals.css";
+import Provider from "../context/provider";
 import { Session } from "next-auth";
 import { Toaster } from "@/components/ui/toaster";
 import { boolean } from "zod";
+import NavBar from "../components/NavBar";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
@@ -24,8 +24,8 @@ export default function RootLayout({
     <html lang="en">
       <body className={inter.className}>
         <Provider session={session}>
+          <NavBar />
           <main>
-            <NavBar />
             <div id="bg-grid">
               <div id="blur-grid"></div>
             </div>
