+++ src/app/(dynamicPage)/[userId]/skeleton.tsx
